git-svn-id: http://svn.reddes.bvsalud.org/scielo-metodologia/trunk@3661 7b3c32f5-e272-4d6f-9936-548b95b19007

+++ www/proc/scielo_crs/shs/xref_reseting.sh
@@ -1,4 +1,16 @@
 . crossref_config.sh
 
-rm $XREF_DB_PATH/*.* 
-rm -rf ../output/crossref/
+if [ "$1" = "çlkjiuy" ]
+then
+	rm $XREF_DB_PATH/*.* 
+	if [ -d ../output/crossref/ ]
+	then
+		rm -rf ../output/crossref/*
+	fi
+	if [ -d ../databases/budget/ ]
+	then
+		rm -rf ../databases/budget/*
+	fi
+	echo reseted
+fi
+
